+++ resources/scripts/fileSystem/fileChecker.js
@@ -2,19 +2,19 @@ const fs = require('fs');
 
 class FileChecker{
     static isApexClass(filePath){
-        return filePath.indexOf('.cls') != -1;
+        return filePath.endsWith('.cls');
     }
     static isApexTrigger(filePath){
-        return filePath.indexOf('.trigger') != -1;
+        return filePath.endsWith('.trigger');
     }
     static isJavaScript(filePath){
-        return filePath.indexOf('.js') != -1;
+        return filePath.endsWith('.js');
     }
     static isAuraDoc(filePath){
-        return filePath.indexOf('.auradoc') != -1;
+        return filePath.endsWith('.auradoc');
     }
     static isAuraComponent(filePath){
-        return filePath.indexOf('.cmp') != -1;
+        return filePath.endsWith('.cmp');
     }
     static isAuraComponentFolder(filePath){
         return filePath.indexOf('/aura/') !== -1 || filePath.indexOf('\\aura\\') !== -1;